# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Cloud Novels API",
    "description": "小说文本转语音服务 — 基于 Fastify + Zod + 阿里云百炼 CosyVoice",
    "version": "0.1.0"
  },
...
  "paths": {
    "/novel/upload": {
      "post": {
        "summary": "上传文本",
        "tags": [
          "novel"
        ],
        "description": "上传小说原始文本，自动解析为章节列表",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "novel_title",
                  "content"
                ],
                "properties": {
                  "novel_title": {
                    "type": "string",
                    "description": "小说名称"
                  },
                  "content": {
                    "type": "string",
                    "description": "小说文本内容"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "解析成功",
            "content": {
              "application/json": {
                "schema": {
                  "description": "解析成功",
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "description": "业务状态码，0 表示成功"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "novel_title": {
                          "type": "string"
                        },
                        "chapters": {
                          "type": "object",
                          "properties": {
                            "total": {
                              "type": "integer"
                            },
                            "pageNum": {
                              "type": "integer"
                            },
                            "pageSize": {
                              "type": "integer"
                            },
                            "list": {
                              "type": "array",
                              "items": {
                                "type": "object"
                              }
                            }
                          }
                        }
                      }
                    },
                    "errorMsg": {
                      "type": "string",
                      "nullable": true,
                      "description": "错误信息，成功时为 null"
                    },
                    "timestamp": {
                      "type": "string",
                      "description": "响应时间"
                    },
                    "traceId": {
                      "type": "string",
                      "description": "请求追踪 ID"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/novel/analyze": {
      "post": {
        "summary": "角色分析",
        "tags": [
          "novel"
        ],
        "description": "对已有章节进行 LLM 角色分析，不含语音合成",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "novel_title",
                  "chapters"
                ],
                "properties": {
                  "novel_title": {
                    "type": "string",
                    "description": "小说名称"
                  },
                  "chapters": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    },
                    "description": "章节列表"
                  },
                  "character_descriptions": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "string"
                    },
                    "description": "可选的角色声音描述"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "分析完成",
            "content": {
              "application/json": {
                "schema": {
                  "description": "分析完成",
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "description": "业务状态码，0 表示成功"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "characters": {
                          "type": "object"
                        }
                      }
                    },
                    "errorMsg": {
                      "type": "string",
                      "nullable": true,
                      "description": "错误信息，成功时为 null"
                    },
                    "timestamp": {
                      "type": "string",
                      "description": "响应时间"
                    },
                    "traceId": {
                      "type": "string",
                      "description": "请求追踪 ID"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/novel/convert": {
      "post": {
        "summary": "合成语音",
        "tags": [
          "novel"
        ],
        "description": "完整合成流程：角色分析 → 注册声音 → 逐句语音合成",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "novel_title",
                  "chapters"
                ],
                "properties": {
                  "novel_title": {
                    "type": "string",
                    "description": "小说名称"
                  },
                  "chapters": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    },
                    "description": "章节列表"
                  },
                  "output_format": {
                    "type": "string",
                    "enum": [
                      "wav",
                      "mp3",
                      "pcm"
                    ],
                    "default": "mp3"
                  },
                  "merge": {
                    "type": "boolean",
                    "default": false,
                    "description": "是否合并为单个音频文件"
                  },
                  "cache": {
                    "type": "boolean",
                    "default": true,
                    "description": "是否使用音频缓存"
                  },
                  "character_descriptions": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "string"
                    },
                    "description": "角色声音描述"
                  },
                  "character_overrides": {
                    "type": "object",
                    "description": "角色画像覆盖"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "合成任务创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "description": "合成任务创建成功",
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "description": "业务状态码，0 表示成功"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "task_id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "task_status": {
                          "type": "string",
                          "enum": [
                            "completed",
                            "processing"
                          ]
                        },
                        "novel_id": {
                          "type": "string"
                        },
                        "chapters": {
                          "type": "object"
                        },
                        "characters_registered": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "merged_url": {
                          "type": "string"
                        }
                      }
                    },
                    "errorMsg": {
                      "type": "string",
                      "nullable": true,
                      "description": "错误信息，成功时为 null"
                    },
                    "timestamp": {
                      "type": "string",
                      "description": "响应时间"
                    },
                    "traceId": {
                      "type": "string",
                      "description": "请求追踪 ID"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/novel/audio": {
      "get": {
        "summary": "音频缓存查询",
        "tags": [
          "novel"
        ],
        "description": "查询小说的音频缓存记录",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "novel_id",
            "required": false,
            "description": "小说 ID"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "novel_title",
            "required": false,
            "description": "小说名称"
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "description": "查询成功",
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "description": "业务状态码，0 表示成功"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "audio": {
                          "type": "object"
                        }
                      }
                    },
                    "errorMsg": {
                      "type": "string",
                      "nullable": true,
                      "description": "错误信息，成功时为 null"
                    },
                    "timestamp": {
                      "type": "string",
                      "description": "响应时间"
                    },
                    "traceId": {
                      "type": "string",
                      "description": "请求追踪 ID"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "开发服务器"
    }
  ],
...
}